// frontends/api/src/generated/v1/models/PaginatedUserListRelationshipList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MIT Learn API
 * MIT public API
 *
 * The version of the OpenAPI document: 0.0.1 (v1)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime"
import type { UserListRelationship } from "./UserListRelationship"
import {
  UserListRelationshipFromJSON,
  UserListRelationshipFromJSONTyped,
  UserListRelationshipToJSON,
  UserListRelationshipToJSONTyped,
} from "./UserListRelationship"

/**
 *
 * @export
 * @interface PaginatedUserListRelationshipList
 */
export interface PaginatedUserListRelationshipList {
  /**
   *
   * @type {number}
   * @memberof PaginatedUserListRelationshipList
   */
  count: number
  /**
   *
   * @type {string}
   * @memberof PaginatedUserListRelationshipList
   */
  next?: string | null
  /**
   *
   * @type {string}
   * @memberof PaginatedUserListRelationshipList
   */
  previous?: string | null
  /**
   *
   * @type {Array<UserListRelationship>}
   * @memberof PaginatedUserListRelationshipList
   */
  results: Array<UserListRelationship>
}

/**
 * Check if a given object implements the PaginatedUserListRelationshipList interface.
 */
export function instanceOfPaginatedUserListRelationshipList(
  value: object,
): value is PaginatedUserListRelationshipList {
  if (!("count" in value) || value["count"] === undefined) return false
  if (!("results" in value) || value["results"] === undefined) return false
  return true
}

export function PaginatedUserListRelationshipListFromJSON(
  json: any,
): PaginatedUserListRelationshipList {
  return PaginatedUserListRelationshipListFromJSONTyped(json, false)
}

export function PaginatedUserListRelationshipListFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): PaginatedUserListRelationshipList {
  if (json == null) {
    return json
  }
  return {
    count: json["count"],
    next: json["next"] == null ? undefined : json["next"],
    previous: json["previous"] == null ? undefined : json["previous"],
    results: (json["results"] as Array<any>).map(UserListRelationshipFromJSON),
  }
}

export function PaginatedUserListRelationshipListToJSON(
  json: any,
): PaginatedUserListRelationshipList {
  return PaginatedUserListRelationshipListToJSONTyped(json, false)
}

export function PaginatedUserListRelationshipListToJSONTyped(
  value?: PaginatedUserListRelationshipList | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    count: value["count"],
    next: value["next"],
    previous: value["previous"],
    results: (value["results"] as Array<any>).map(UserListRelationshipToJSON),
  }
}