// frontends/api/src/generated/v1/models/UserListRelationship.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MIT Learn API
 * MIT public API
 *
 * The version of the OpenAPI document: 0.0.1 (v1)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime"
import type { LearningResource } from "./LearningResource"
import {
  LearningResourceFromJSON,
  LearningResourceFromJSONTyped,
  LearningResourceToJSON,
  LearningResourceToJSONTyped,
} from "./LearningResource"

/**
 * Serializer for UserListRelationship model
 * @export
 * @interface UserListRelationship
 */
export interface UserListRelationship {
  /**
   *
   * @type {number}
   * @memberof UserListRelationship
   */
  readonly id: number
  /**
   *
   * @type {LearningResource}
   * @memberof UserListRelationship
   */
  readonly resource: LearningResource
  /**
   *
   * @type {number}
   * @memberof UserListRelationship
   */
  position?: number
  /**
   *
   * @type {number}
   * @memberof UserListRelationship
   */
  parent: number
  /**
   *
   * @type {number}
   * @memberof UserListRelationship
   */
  child: number
}

/**
 * Check if a given object implements the UserListRelationship interface.
 */
export function instanceOfUserListRelationship(
  value: object,
): value is UserListRelationship {
  if (!("id" in value) || value["id"] === undefined) return false
  if (!("resource" in value) || value["resource"] === undefined) return false
  if (!("parent" in value) || value["parent"] === undefined) return false
  if (!("child" in value) || value["child"] === undefined) return false
  return true
}

export function UserListRelationshipFromJSON(json: any): UserListRelationship {
  return UserListRelationshipFromJSONTyped(json, false)
}

export function UserListRelationshipFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): UserListRelationship {
  if (json == null) {
    return json
  }
  return {
    id: json["id"],
    resource: LearningResourceFromJSON(json["resource"]),
    position: json["position"] == null ? undefined : json["position"],
    parent: json["parent"],
    child: json["child"],
  }
}

export function UserListRelationshipToJSON(json: any): UserListRelationship {
  return UserListRelationshipToJSONTyped(json, false)
}

export function UserListRelationshipToJSONTyped(
  value?: Omit<UserListRelationship, "id" | "resource"> | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    position: value["position"],
    parent: value["parent"],
    child: value["child"],
  }
}