// frontends/api/src/generated/v1/models/MicroUserListRelationship.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MIT Learn API
 * MIT public API
 *
 * The version of the OpenAPI document: 0.0.1 (v1)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime"
/**
 * Serializer containing only parent and child ids for a user list relationship
 * @export
 * @interface MicroUserListRelationship
 */
export interface MicroUserListRelationship {
  /**
   *
   * @type {number}
   * @memberof MicroUserListRelationship
   */
  readonly id: number
  /**
   * The id of the parent learning resource
   * @type {number}
   * @memberof MicroUserListRelationship
   */
  readonly parent: number
  /**
   *
   * @type {number}
   * @memberof MicroUserListRelationship
   */
  readonly child: number
}

/**
 * Check if a given object implements the MicroUserListRelationship interface.
 */
export function instanceOfMicroUserListRelationship(
  value: object,
): value is MicroUserListRelationship {
  if (!("id" in value) || value["id"] === undefined) return false
  if (!("parent" in value) || value["parent"] === undefined) return false
  if (!("child" in value) || value["child"] === undefined) return false
  return true
}

export function MicroUserListRelationshipFromJSON(
  json: any,
): MicroUserListRelationship {
  return MicroUserListRelationshipFromJSONTyped(json, false)
}

export function MicroUserListRelationshipFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): MicroUserListRelationship {
  if (json == null) {
    return json
  }
  return {
    id: json["id"],
    parent: json["parent"],
    child: json["child"],
  }
}

export function MicroUserListRelationshipToJSON(
  json: any,
): MicroUserListRelationship {
  return MicroUserListRelationshipToJSONTyped(json, false)
}

export function MicroUserListRelationshipToJSONTyped(
  value?: Omit<MicroUserListRelationship, "id" | "parent" | "child"> | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {}
}